fix(l1): fix replay with newest ethrex changes (#33)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Because of the functionality for executing blocks with add_block replay
was tightly coupled with ethrex code, specially the In Memory database,
we had to access inner fields of it. The structure of the database has
changed and what we used to do is now more complex and also it's hard to
maintain, we don't want replay to break and we want to be flexible.
I think that it's time for us to just implement the replay feature
inside ethrex code in a few lines, this will remove the need of having
the --no-zkvm flag because now we will be able to run ANY block with
zkVMs, thanks to the modified MPT. Other reason add_block was useful was
for running a samply, but the things that matter the most in the samply
are executing the block and applying account updates, and the stateless
execution does exactly the same, the only difference are some
validations but these are not significant in terms of performance.

- Remove support for `add_block` by adding a replay feature in the trie
crate of ethrex

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number