feat(l1): support replaying custom L1 blocks with ETH transfers (#41)
**Description**

- Adds transaction builder module with code for building signed ETH
transfer and ERC20 transfer transactions.
- Adds flags `--n-txs` to pass the number of transactions to include in
the custom block being executed and `--tx` with options `eth-transfer` &
`erc20-transfer` to specify the tx kind to include in the custom block
(for now, only 1 tx kind per block).
- ERC20 transfers are not yet fully-supported though, their support will
be achieved in another PR because the setup is different and more
complex since we need to include a deploy tx and one or multiple
allowance txs to transfer the tokens.